style: move schema in entities
