+++ src/entities/schemas/index.ts
@@ -1,2 +1,3 @@
 export { Client } from './client';
 export { getAuthSchema } from './auth';
+export { VariablesSchema } from './variables';
